StressTest/requirements.txt
astroid==2.4.2
asyncio==3.4.3
attrs==20.2.0
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
hashids==1.3.1
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.0.1
isort==5.5.3
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.5.0
packaging==20.4
paramiko==2.7.2
pluggy==0.13.1
py==1.9.0
pycparser==2.20
pylint==2.6.0
PyMySQL==0.10.1
PyNaCl==1.4.0
pyparsing==2.4.7
pytest==6.1.0
pytest-asyncio==0.14.0
requests==2.24.0
six==1.15.0
sshtunnel==0.1.5
toml==0.10.1
tornado==6.0.4
typed-ast==1.4.1
urllib3==1.25.10
websocket-client==0.57.0
websockets==8.1
wrapt==1.12.1
zipp==3.2.0
